(refactor) folder structure / pages naming
