+++ src/pages/ConditionsList/index.tsx
